backend/requirements.txt
flask
paho-mqtt
pymongo